LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/position-try-container-query-expected.txt


PASS Size container query responds to fallback width
FAIL Size container query responds to fallback width and applies height to not fit the first fallback assert_equals: expected "0px" but got "100px"
